refactor:  change namespace using keuangan and some typos
