+++ app/Support/Menu.php
@@ -105,6 +105,7 @@ class Menu
                     'keuangan.posting-jurnal',
                     'keuangan.laporan-faktur-pajak.read',
                     'keuangan.igd-ke-rawat-inap.read',
+                    'keuangan.obat-ralan-ke-ranap.read'
                 ]),
                 'items' => [
                     [
@@ -275,6 +276,12 @@ class Menu
                         'type'              => 'link',
                         'hasAnyPermissions' => $user->can('keuangan.igd-ke-rawat-inap.read'),
                     ],
+                    [
+                        'name'              => 'Obat Ralan ke Ranap',
+                        'icon'              => 'fas fa-exchange-alt',
+                        'url'               => route('admin.keuangan.obat-ralan-ke-ranap'),
+                        'hasAnyPermissions' => $user->can('keuangan.obat-ralan-ke-ranap.read'),
+                    ],
                 ],
             ],
             [
@@ -296,7 +303,6 @@ class Menu
                     'farmasi.defecta-depo.read',
                     'farmasi.daftar-riwayat-obat-alkes.read',
                     'farmasi.farmasi.rincian-perbandingan-po.read',
-                    'farmasi.obat-ralan-ke-ranap.read'
                 ]),
                 'items' => [
                     [
@@ -388,12 +394,6 @@ class Menu
                         'icon'              => 'fas fa-balance-scale',
                         'url'               => route('admin.farmasi.rincian-kunjungan-ralan'),
                         'hasAnyPermissions' => $user->can('farmasi.rincian-kunjungan-ralan.read'),
-                    ],
-                    [
-                        'name'              => 'Obat Rawat Jalan ke Rawat Inap',
-                        'icon'              => 'fas fa-exchange-alt',
-                        'url'               => route('admin.farmasi.obat-ralan-ke-ranap'),
-                        'hasAnyPermissions' => $user->can('farmasi.obat-ralan-ke-ranap.read'),
                     ],
                 ],
             ],
